fix: clear the seven error-level clang-tidy findings
They predate this branch -- the same seven come out of main -- and nobody
had seen them: the analysis half of the gate is opt-in, and the hook that
would have asked for it pointed at a directory that no longer existed
until the previous commit repaired it. So the gate has been half red for
as long as the analyser has been wired up, and repairing the hook is what
made that matter: a push to main from a terminal now asks, and a yes now
blocks.

PTPMaster's LogIntervalToNs took an int8_t and shifted an int64 by it.
MsToLogInterval clamps its result to -128 and 127 rather than to what the
shift can hold, so 127 is representable and a shift of 127 is undefined
behaviour -- not a large number, undefined. The reachable range really is
what the comment claimed, at most 21, because the input is a positive int
of milliseconds; but a comment is not a bound. The shifts are now clamped
to 33 and 63, the last values an int64 holds with kNsPerSecond, so the
limit sits where the compiler can see it. Nothing reachable changes.

The other five are multiplications performed in int or uint32_t and
widened to size_t afterwards, which is the wrong order: the overflow, if
it ever came, would happen before the widening rather than after it. Each
now multiplies in size_t.

Verified: scripts/ci-local.sh with AES67_ANALYSE=1 PASS -- 42/42 tests and
112 files analysed with no error-level findings, the first time the full
gate has passed.

+++ packages/aes67_macos_driver/Examples/SimpleSDPParse.cpp
@@ -80,7 +80,12 @@ int main(int argc, char* argv[]) {
     printSDPSession(*session);
 
     // Calculate some useful information
-    size_t samplesPerSecond = session->sampleRate * session->numChannels;
+    // The multiplication is done in size_t, not in uint32_t and widened
+    // afterwards: at 192 kHz and 128 channels the product still fits in 32
+    // bits, but nothing here says it has to, and the widening happens after
+    // the overflow rather than before it.
+    size_t samplesPerSecond =
+        static_cast<size_t>(session->sampleRate) * session->numChannels;
     size_t bytesPerSecond;
 
     if (session->encoding == "L24") {
